requirements.txt
python-telegram-bot==21.6
Pillow>=10.0.0
python-dotenv>=1.0.0